refactor(admin): profile workspace residual 분리
