+++ front/src/routes/Admin/AdminProfileWorkspacePageModel.ts
@@ -1,70 +1,40 @@
 import { type DehydratedState, useQueryClient } from "@tanstack/react-query"
 import { useRouter } from "next/router"
-import { ChangeEvent, useCallback, useEffect, useMemo, useRef, useState } from "react"
+import { useCallback, useEffect, useMemo, useState } from "react"
 import { apiFetch } from "src/apis/backend/client"
-import type { ProfileCardLinkItem } from "src/constants/profileCardLinks"
 import type { AuthMember } from "src/hooks/useAuthSession"
 import useAuthSession from "src/hooks/useAuthSession"
 import { setAdminProfileCache, toAdminProfile } from "src/hooks/useAdminProfile"
 import { setProfileWorkspaceCache, useProfileWorkspace } from "src/hooks/useProfileWorkspace"
-import useViewportImageEditor from "src/libs/imageEditor/useViewportImageEditor"
 import {
   buildProfileWorkspaceAdminProfileCacheFields,
-  AboutProjectBlock,
   normalizeProfileWorkspaceContent,
   ProfileWorkspaceContent,
   ProfileWorkspaceResponse,
   serializeProfileWorkspaceContent,
-  AboutSectionBlock,
 } from "src/libs/profileWorkspace"
-import {
-  buildImageOptimizationSummary,
-  buildProfileImageEditedFile,
-  clampProfileImageEditFocusBySource,
-  clampProfileImageEditZoom,
-  normalizeProfileImageUploadError,
-  prepareProfileImageForUpload,
-  ProfileImageSourceSize,
-  PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_X,
-  PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_Y,
-  PROFILE_IMAGE_EDIT_MIN_ZOOM,
-  resolveProfileImageEditDrawRatios,
-} from "src/libs/profileImageUpload"
 import { saveProfileCardWithConflictRetry } from "src/libs/profileCardSave"
-import { acquireBodyScrollLock } from "src/libs/utils/bodyScrollLock"
 import {
   WORKSPACE_SECTIONS,
   buildWorkspaceFallback,
-  createBlankAboutProject,
-  createBlankAboutSection,
-  createBlankLinkItem,
   type LinkTab,
-  moveListItem,
   type PreviewMode,
-  reorderListItem,
   serializeWorkspaceSection,
   toPayloadLinks,
   validateLinkInputs,
   type WorkspaceSectionId,
 } from "src/routes/Admin/AdminProfileWorkspaceModel"
 import {
-  PROFILE_IMAGE_DRAFT_DEFAULT_SOURCE_SIZE,
-  PROFILE_IMAGE_UPLOAD_RETRY_DELAY_MS,
   PROFILE_UNSAVED_CHANGES_MESSAGE,
-  parseResponseErrorBody,
-  readImageSourceSizeFromFile,
-  requestProfileImageUpload,
   revalidatePublicBlogAppearance,
-  sleep,
 } from "src/routes/Admin/AdminProfilePersistenceModel"
+import {
+  type OpenIconPicker,
+  useAdminProfileWorkspaceDraftActions,
+} from "src/routes/Admin/AdminProfileWorkspacePageDraftActions"
+import { useAdminProfileWorkspaceImageDraft } from "src/routes/Admin/AdminProfileWorkspacePageImageDraft"
 
 export type NoticeTone = "idle" | "loading" | "success" | "error"
-type OpenIconPicker = `${LinkTab}:${number}` | null
-type ProfileImageDraftTransformState = {
-  focusX: number
-  focusY: number
-  zoom: number
-}
 
 export type AdminProfileWorkspacePageProps = {
   dehydratedState: DehydratedState
@@ -115,23 +85,29 @@ export const useAdminProfileWorkspacePageModel = ({
   const [remoteDraft, setRemoteDraft] = useState<ProfileWorkspaceContent>(fallbackWorkspace.draft)
   const [publishedSnapshot, setPublishedSnapshot] = useState<ProfileWorkspaceContent>(fallbackWorkspace.published)
   const [draft, setDraft] = useState<ProfileWorkspaceContent>(fallbackWorkspace.draft)
-  const [profileImageFileName, setProfileImageFileName] = useState("")
-  const [isProfileImageEditorOpen, setIsProfileImageEditorOpen] = useState(false)
-  const [profileImageDraftFile, setProfileImageDraftFile] = useState<File | null>(null)
-  const [profileImageDraftPreviewUrl, setProfileImageDraftPreviewUrl] = useState("")
-  const [profileImageDraftFocusX, setProfileImageDraftFocusX] = useState(PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_X)
-  const [profileImageDraftFocusY, setProfileImageDraftFocusY] = useState(PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_Y)
-  const [profileImageDraftZoom, setProfileImageDraftZoom] = useState(PROFILE_IMAGE_EDIT_MIN_ZOOM)
-  const [profileImageDraftSourceSize, setProfileImageDraftSourceSize] = useState<ProfileImageSourceSize>(
-    PROFILE_IMAGE_DRAFT_DEFAULT_SOURCE_SIZE
-  )
-  const [profileImageDraftNotice, setProfileImageDraftNotice] = useState<{ tone: NoticeTone; text: string }>({
-    tone: "idle",
-    text: "",
+  const {
+    addAboutItem,
+    addAboutProject,
+    addAboutSection,
+    appendLinkItem,
+    moveAboutItem,
+    moveAboutProject,
+    moveAboutSection,
+    moveLinkItem,
+    removeAboutItem,
+    removeAboutProject,
+    removeAboutSection,
+    removeLinkItem,
+    reorderLinkItems,
+    updateAboutProject,
+    updateAboutSection,
+    updateDraft,
+    updateLinkItem,
+  } = useAdminProfileWorkspaceDraftActions({
+    setDraft,
+    setOpenIconPicker,
   })
-  const profileImageDraftFrameRef = useRef<HTMLDivElement>(null)
-  const profileImageFileInputRef = useRef<HTMLInputElement>(null)
-  const profileImageDraftFileSeqRef = useRef(0)
+
 
   const syncPublishedAdminProfileCache = useCallback(
     (content: ProfileWorkspaceContent) => {
@@ -214,21 +190,7 @@ export const useAdminProfileWorkspacePageModel = ({
     }
   }, [])
 
-  useEffect(() => {
-    return () => {
-      if (profileImageDraftPreviewUrl) {
-        URL.revokeObjectURL(profileImageDraftPreviewUrl)
-      }
-    }
-  }, [profileImageDraftPreviewUrl])
 
-  useEffect(() => {
-    if (!isProfileImageEditorOpen) return
-    const releaseBodyScrollLock = acquireBodyScrollLock()
-    return () => {
-      releaseBodyScrollLock()
-    }
-  }, [isProfileImageEditorOpen])
 
   const hasWorkspaceUnsavedChanges = useMemo(
     () => serializeProfileWorkspaceContent(draft) !== serializeProfileWorkspaceContent(remoteDraft),
@@ -308,6 +270,52 @@ export const useAdminProfileWorkspacePageModel = ({
     [applyWorkspaceState]
   )
 
+  const {
+    applyProfileImageDraftPreviewStyle,
+    clearProfileImageDraft,
+    commitProfileImageDraftTransform,
+    computeAnchoredZoomTransform,
+    computeDraggedProfileImageTransform,
+    finalizeProfileImageDraftPointer,
+    handleApplyProfileImageDraft,
+    handleDraftFileChange,
+    handleProfileImageDraftPointerDown,
+    handleProfileImageDraftPointerMove,
+    handleUploadMemberProfileImage,
+    isProfileImageDraftDragging,
+    isProfileImageEditorOpen,
+    normalizeProfileImageDraftTransform,
+    profileImageDraftFile,
+    profileImageDraftFileSeqRef,
+    profileImageDraftFocusX,
+    profileImageDraftFocusY,
+    profileImageDraftFrameRef,
+    profileImageDraftNotice,
+    profileImageDraftPreviewUrl,
+    profileImageDraftSourceSize,
+    profileImageDraftTransformRef,
+    profileImageDraftZoom,
+    profileImageFileInputRef,
+    profileImageFileName,
+    resetProfileImageDraftInteractions,
+    scheduleProfileImageDraftTransform,
+    setIsProfileImageEditorOpen,
+    setProfileImageDraftFile,
+    setProfileImageDraftFocusX,
+    setProfileImageDraftFocusY,
+    setProfileImageDraftNotice,
+    setProfileImageDraftPreviewUrl,
+    setProfileImageDraftSourceSize,
+    setProfileImageDraftZoom,
+    setProfileImageFileName,
+  } = useAdminProfileWorkspaceImageDraft({
+    refreshWorkspace,
+    sessionMemberId: sessionMember?.id,
+    setImageNotice,
+    setLoadingKey,
+    setMe,
+  })
+
   const persistDisplayName = useCallback(
     async (memberId: number, nickname: string) => {
       const updatedMember = await saveProfileCardWithConflictRetry(() =>
@@ -374,426 +382,9 @@ export const useAdminProfileWorkspacePageModel = ({
     )
   }, [buildDraftPayload, sessionMember?.id])
 
-  const updateDraft = useCallback(
-    (
-      field: keyof ProfileWorkspaceContent,
-      value:
-        | string
-        | ProfileCardLinkItem[]
-        | AboutSectionBlock[]
-        | ((current: ProfileWorkspaceContent) => ProfileWorkspaceContent)
-    ) => {
-      if (typeof value === "function") {
-        setDraft((current) => value(current))
-        return
-      }
 
-      setDraft((current) => ({
-        ...current,
-        [field]: value,
-      }))
-    },
-    []
-  )
 
-  const updateLinkItem = useCallback(
-    (section: LinkTab, index: number, field: keyof ProfileCardLinkItem, value: string) => {
-      setDraft((current) => {
-        const key = section === "service" ? "serviceLinks" : "contactLinks"
-        return {
-          ...current,
-          [key]: current[key].map((item, itemIndex) =>
-            itemIndex === index
-              ? {
-                  ...item,
-                  [field]: value,
-                }
-              : item
-          ),
-        }
-      })
-    },
-    []
-  )
 
-  const appendLinkItem = useCallback((section: LinkTab) => {
-    setDraft((current) => {
-      const key = section === "service" ? "serviceLinks" : "contactLinks"
-      return {
-        ...current,
-        [key]: [...current[key], createBlankLinkItem(section)],
-      }
-    })
-  }, [])
-
-  const removeLinkItem = useCallback((section: LinkTab, index: number) => {
-    setDraft((current) => {
-      const key = section === "service" ? "serviceLinks" : "contactLinks"
-      return {
-        ...current,
-        [key]: current[key].filter((_, itemIndex) => itemIndex !== index),
-      }
-    })
-    setOpenIconPicker((current) => (current === `${section}:${index}` ? null : current))
-  }, [])
-
-  const moveLinkItem = useCallback((section: LinkTab, index: number, direction: -1 | 1) => {
-    setDraft((current) => {
-      const key = section === "service" ? "serviceLinks" : "contactLinks"
-      return {
-        ...current,
-        [key]: moveListItem(current[key], index, direction),
-      }
-    })
-  }, [])
-
-  const reorderLinkItems = useCallback((section: LinkTab, fromIndex: number, toIndex: number) => {
-    setDraft((current) => {
-      const key = section === "service" ? "serviceLinks" : "contactLinks"
-      return {
-        ...current,
-        [key]: reorderListItem(current[key], fromIndex, toIndex),
-      }
-    })
-  }, [])
-
-  const updateAboutSection = useCallback((sectionIndex: number, updater: (section: AboutSectionBlock) => AboutSectionBlock) => {
-    setDraft((current) => ({
-      ...current,
-      aboutSections: current.aboutSections.map((section, index) =>
-        index === sectionIndex ? updater(section) : section
-      ),
-    }))
-  }, [])
-
-  const addAboutSection = useCallback(() => {
-    setDraft((current) => ({
-      ...current,
-      aboutSections: [...current.aboutSections, createBlankAboutSection()],
-    }))
-  }, [])
-
-  const removeAboutSection = useCallback((sectionIndex: number) => {
-    setDraft((current) => ({
-      ...current,
-      aboutSections: current.aboutSections.filter((_, index) => index !== sectionIndex),
-    }))
-  }, [])
-
-  const moveAboutSection = useCallback((sectionIndex: number, direction: -1 | 1) => {
-    setDraft((current) => ({
-      ...current,
-      aboutSections: moveListItem(current.aboutSections, sectionIndex, direction),
-    }))
-  }, [])
-
-  const addAboutItem = useCallback((sectionIndex: number) => {
-    updateAboutSection(sectionIndex, (section) => ({
-      ...section,
-      items: [...section.items, ""],
-    }))
-  }, [updateAboutSection])
-
-  const removeAboutItem = useCallback((sectionIndex: number, itemIndex: number) => {
-    updateAboutSection(sectionIndex, (section) => ({
-      ...section,
-      items: section.items.filter((_, index) => index !== itemIndex),
-    }))
-  }, [updateAboutSection])
-
-  const moveAboutItem = useCallback((sectionIndex: number, itemIndex: number, direction: -1 | 1) => {
-    updateAboutSection(sectionIndex, (section) => ({
-      ...section,
-      items: moveListItem(section.items, itemIndex, direction),
-    }))
-  }, [updateAboutSection])
-
-  const updateAboutProject = useCallback((projectIndex: number, updater: (project: AboutProjectBlock) => AboutProjectBlock) => {
-    setDraft((current) => ({
-      ...current,
-      aboutProjects: current.aboutProjects.map((project, index) =>
-        index === projectIndex ? updater(project) : project
-      ),
-    }))
-  }, [])
-
-  const addAboutProject = useCallback(() => {
-    setDraft((current) => ({
-      ...current,
-      aboutProjects: [...current.aboutProjects, createBlankAboutProject()],
-    }))
-  }, [])
-
-  const removeAboutProject = useCallback((projectIndex: number) => {
-    setDraft((current) => ({
-      ...current,
-      aboutProjects: current.aboutProjects.filter((_, index) => index !== projectIndex),
-    }))
-  }, [])
-
-  const moveAboutProject = useCallback((projectIndex: number, direction: -1 | 1) => {
-    setDraft((current) => ({
-      ...current,
-      aboutProjects: moveListItem(current.aboutProjects, projectIndex, direction),
-    }))
-  }, [])
-
-  const applyProfileImageDraftPreviewStyle = useCallback(
-    (transform: ProfileImageDraftTransformState) => {
-      const frame = profileImageDraftFrameRef.current
-      if (!frame) return
-
-      const { drawWidth, drawHeight } = resolveProfileImageEditDrawRatios(
-        profileImageDraftSourceSize,
-        transform.zoom
-      )
-      const centerXRatio = transform.focusX / 100
-      const centerYRatio = transform.focusY / 100
-      const leftRatio = centerXRatio - drawWidth / 2
-      const topRatio = centerYRatio - drawHeight / 2
-
-      frame.style.setProperty("--profile-draft-width", `${drawWidth * 100}%`)
-      frame.style.setProperty("--profile-draft-height", `${drawHeight * 100}%`)
-      frame.style.setProperty("--profile-draft-left", `${leftRatio * 100}%`)
-      frame.style.setProperty("--profile-draft-top", `${topRatio * 100}%`)
-    },
-    [profileImageDraftSourceSize]
-  )
-
-  const normalizeProfileImageDraftTransform = useCallback(
-    (current: ProfileImageDraftTransformState) => {
-      const zoom = clampProfileImageEditZoom(current.zoom)
-      const clampedFocus = clampProfileImageEditFocusBySource({
-        focusX: current.focusX,
-        focusY: current.focusY,
-        zoom,
-        sourceSize: profileImageDraftSourceSize,
-      })
-
-      return {
-        focusX: clampedFocus.focusX,
-        focusY: clampedFocus.focusY,
-        zoom,
-      }
-    },
-    [profileImageDraftSourceSize]
-  )
-
-  const computeAnchoredZoomTransform = useCallback(
-    (
-      baseTransform: ProfileImageDraftTransformState,
-      nextZoom: number,
-      anchorXRatio: number,
-      anchorYRatio: number
-    ): ProfileImageDraftTransformState => {
-      const { drawWidth: prevDrawWidth, drawHeight: prevDrawHeight } = resolveProfileImageEditDrawRatios(
-        profileImageDraftSourceSize,
-        baseTransform.zoom
-      )
-      const { drawWidth: nextDrawWidth, drawHeight: nextDrawHeight } = resolveProfileImageEditDrawRatios(
-        profileImageDraftSourceSize,
-        nextZoom
-      )
-      const prevCenterX = baseTransform.focusX / 100
-      const prevCenterY = baseTransform.focusY / 100
-      const prevLeft = prevCenterX - prevDrawWidth / 2
-      const prevTop = prevCenterY - prevDrawHeight / 2
-      const pointerImageX = Math.min(1, Math.max(0, (anchorXRatio - prevLeft) / prevDrawWidth))
-      const pointerImageY = Math.min(1, Math.max(0, (anchorYRatio - prevTop) / prevDrawHeight))
-      const nextLeft = anchorXRatio - pointerImageX * nextDrawWidth
-      const nextTop = anchorYRatio - pointerImageY * nextDrawHeight
-
-      return {
-        focusX: (nextLeft + nextDrawWidth / 2) * 100,
-        focusY: (nextTop + nextDrawHeight / 2) * 100,
-        zoom: nextZoom,
-      }
-    },
-    [profileImageDraftSourceSize]
-  )
-
-  const computeDraggedProfileImageTransform = useCallback(
-    (baseTransform: ProfileImageDraftTransformState, deltaXRatio: number, deltaYRatio: number) => {
-      const zoomScale = Math.max(baseTransform.zoom, PROFILE_IMAGE_EDIT_MIN_ZOOM)
-      return {
-        focusX: baseTransform.focusX + deltaXRatio * (100 / zoomScale),
-        focusY: baseTransform.focusY + deltaYRatio * (100 / zoomScale),
-        zoom: baseTransform.zoom,
-      }
-    },
-    []
-  )
-
-  const commitProfileImageDraftTransform = useCallback(
-    (normalized: ProfileImageDraftTransformState) => {
-      applyProfileImageDraftPreviewStyle(normalized)
-      setProfileImageDraftFocusX((prev) => (Math.abs(prev - normalized.focusX) > 0.0001 ? normalized.focusX : prev))
-      setProfileImageDraftFocusY((prev) => (Math.abs(prev - normalized.focusY) > 0.0001 ? normalized.focusY : prev))
-      setProfileImageDraftZoom((prev) => (Math.abs(prev - normalized.zoom) > 0.0001 ? normalized.zoom : prev))
-    },
-    [applyProfileImageDraftPreviewStyle]
-  )
-
-  const {
-    commitTransform: commitProfileImageDraftViewportTransform,
-    finalizePointer: finalizeProfileImageDraftPointer,
-    handlePointerDown: handleProfileImageDraftPointerDown,
-    handlePointerMove: handleProfileImageDraftPointerMove,
-    isDragging: isProfileImageDraftDragging,
-    resetInteractions: resetProfileImageDraftInteractions,
-    scheduleTransform: scheduleProfileImageDraftTransform,
-    transformRef: profileImageDraftTransformRef,
-  } = useViewportImageEditor<ProfileImageDraftTransformState>({
-    frameRef: profileImageDraftFrameRef,
-    initialTransform: {
-      focusX: profileImageDraftFocusX,
-      focusY: profileImageDraftFocusY,
-      zoom: profileImageDraftZoom,
-    },
-    enabled: Boolean(profileImageDraftFile),
-    clampZoom: clampProfileImageEditZoom,
-    normalizeTransform: normalizeProfileImageDraftTransform,
-    computeAnchoredZoomTransform,
-    computeDraggedTransform: computeDraggedProfileImageTransform,
-    onCommit: commitProfileImageDraftTransform,
-  })
-
-  const clearProfileImageDraft = useCallback(() => {
-    profileImageDraftFileSeqRef.current += 1
-    resetProfileImageDraftInteractions()
-    setProfileImageDraftFile(null)
-    setProfileImageDraftSourceSize(PROFILE_IMAGE_DRAFT_DEFAULT_SOURCE_SIZE)
-    setProfileImageDraftNotice({ tone: "idle", text: "" })
-    commitProfileImageDraftViewportTransform({
-      focusX: PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_X,
-      focusY: PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_Y,
-      zoom: PROFILE_IMAGE_EDIT_MIN_ZOOM,
-    })
-    setProfileImageDraftPreviewUrl((prev) => {
-      if (prev) URL.revokeObjectURL(prev)
-      return ""
-    })
-  }, [commitProfileImageDraftViewportTransform, resetProfileImageDraftInteractions])
-
-  const handleDraftFileChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0]
-      event.target.value = ""
-      if (!file) return
-
-      const nextFileSeq = profileImageDraftFileSeqRef.current + 1
-      profileImageDraftFileSeqRef.current = nextFileSeq
-      setProfileImageFileName(file.name)
-      setProfileImageDraftFile(file)
-      setProfileImageDraftSourceSize(PROFILE_IMAGE_DRAFT_DEFAULT_SOURCE_SIZE)
-      setProfileImageDraftNotice({ tone: "idle", text: "" })
-      commitProfileImageDraftViewportTransform({
-        focusX: PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_X,
-        focusY: PROFILE_IMAGE_EDIT_DEFAULT_FOCUS_Y,
-        zoom: PROFILE_IMAGE_EDIT_MIN_ZOOM,
-      })
-      setProfileImageDraftPreviewUrl((prev) => {
-        if (prev) URL.revokeObjectURL(prev)
-        return URL.createObjectURL(file)
-      })
-      void readImageSourceSizeFromFile(file)
-        .then((sourceSize) => {
-          if (profileImageDraftFileSeqRef.current !== nextFileSeq) return
-          setProfileImageDraftSourceSize(sourceSize)
-          scheduleProfileImageDraftTransform(profileImageDraftTransformRef.current)
-        })
-        .catch(() => {
-          if (profileImageDraftFileSeqRef.current !== nextFileSeq) return
-          setProfileImageDraftNotice({ tone: "error", text: "이미지 해상도 정보를 읽지 못했습니다." })
-        })
-    },
-    [commitProfileImageDraftViewportTransform, profileImageDraftTransformRef, scheduleProfileImageDraftTransform]
-  )
-
-  useEffect(() => {
-    scheduleProfileImageDraftTransform(profileImageDraftTransformRef.current)
-  }, [profileImageDraftSourceSize, profileImageDraftTransformRef, scheduleProfileImageDraftTransform])
-
-  const handleUploadMemberProfileImage = useCallback(
-    async (selectedFile?: File): Promise<boolean> => {
-      const file = selectedFile || profileImageFileInputRef.current?.files?.[0]
-      const memberId = sessionMember?.id
-      if (!file || !memberId) return false
-
-      try {
-        setLoadingKey("upload")
-        setImageNotice({ tone: "loading", text: "프로필 이미지를 최적화하고 초안에 반영하고 있습니다..." })
-        const prepared = await prepareProfileImageForUpload(file)
-        let uploadResponse = await requestProfileImageUpload(memberId, prepared.file)
-
-        if (uploadResponse.status === 409) {
-          const firstConflictBody = await parseResponseErrorBody(uploadResponse)
-          const retryMessage = "요청 충돌을 감지해 자동 재시도 중입니다..."
-          setImageNotice({ tone: "loading", text: retryMessage })
-          setProfileImageDraftNotice({ tone: "loading", text: retryMessage })
-          await sleep(PROFILE_IMAGE_UPLOAD_RETRY_DELAY_MS)
-          uploadResponse = await requestProfileImageUpload(memberId, prepared.file)
-          if (!uploadResponse.ok) {
-            const retryBody = await parseResponseErrorBody(uploadResponse)
-            throw new Error(`이미지 업로드 실패 (${uploadResponse.status}) ${retryBody || firstConflictBody}`.trim())
-          }
-        } else if (!uploadResponse.ok) {
-          const body = await parseResponseErrorBody(uploadResponse)
-          throw new Error(`이미지 업로드 실패 (${uploadResponse.status}) ${body}`.trim())
-        }
-
-        const uploadData = (await uploadResponse.json()) as AuthMember
-        setMe(uploadData)
-        await refreshWorkspace(memberId)
-        const successMessage = `프로필 이미지가 초안에 반영되었습니다. ${buildImageOptimizationSummary(prepared)}`
-        setImageNotice({ tone: "success", text: successMessage })
-        setProfileImageDraftNotice({ tone: "success", text: successMessage })
-        return true
-      } catch (error) {
-        const message = normalizeProfileImageUploadError(error)
-        setImageNotice({ tone: "error", text: `프로필 이미지 저장 실패: ${message}` })
-        setProfileImageDraftNotice({ tone: "error", text: `프로필 이미지 저장 실패: ${message}` })
-        return false
-      } finally {
-        if (profileImageFileInputRef.current) {
-          profileImageFileInputRef.current.value = ""
-        }
-        setLoadingKey("")
-      }
-    },
-    [refreshWorkspace, sessionMember?.id, setMe]
-  )
-
-  const handleApplyProfileImageDraft = useCallback(async () => {
-    if (!profileImageDraftFile) {
-      setProfileImageDraftNotice({ tone: "error", text: "먼저 프로필 이미지를 선택해주세요." })
-      return
-    }
-
-    try {
-      setProfileImageDraftNotice({ tone: "loading", text: "편집 결과를 업로드하고 있습니다..." })
-      const editedFile = await buildProfileImageEditedFile(profileImageDraftFile, {
-        focusX: profileImageDraftFocusX,
-        focusY: profileImageDraftFocusY,
-        zoom: profileImageDraftZoom,
-      })
-      const uploaded = await handleUploadMemberProfileImage(editedFile)
-      if (uploaded) {
-        setIsProfileImageEditorOpen(false)
-        clearProfileImageDraft()
-      }
-    } catch (error) {
-      const message = normalizeProfileImageUploadError(error)
-      setProfileImageDraftNotice({ tone: "error", text: message })
-    }
-  }, [
-    clearProfileImageDraft,
-    handleUploadMemberProfileImage,
-    profileImageDraftFile,
-    profileImageDraftFocusX,
-    profileImageDraftFocusY,
-    profileImageDraftZoom,
-  ])
 
   const handleSaveDraft = useCallback(async () => {
     if (!sessionMember?.id) return
